// src/app/models/flight-details-itinerary.ts
export declare interface FlightDetailsItinerary {
  carrierName: string;
  departureDate?: string;
  flightNumber?: string;
  cabinType?: string;
  aircraft?: string;
  departureCode?: string;
  arrivalCode?: string;
  arrivalDate?: string
  flightDuration?: string
  checkInStartDate?: string;
  checkInEndDate?: string;
  flightStatus?: string;
}